pyproject.toml
[tool.poetry]
name = "verifiable-tg-bot"
version = "0.1.0"
description = ""
authors = ["Gabriel Fior <fior.gbm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
httpx = "<=0.26.0"
giza-cli = ">=0.10.0"
giza-osiris = "^0.2.5"
#onnx2json = "^2.0.4"
#onnx = "^1.15.0"
#json2onnx = "^2.0.3"
#sktime = "^0.26.0"
#llvmlite = "^0.42.0"
#seaborn = "^0.13.2"
#skl2onnx = "^1.16.0"
giza-actions = "0.1.3" # update for scheduling
python-telegram-bot = "20.7"
pandas = "^2.2.0"
giza-datasets = "^0.1.1"
python-dotenv = "^1.0.1"
giza = "^0.5.18"
pytelegrambotapi = "^4.15.4"



[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
